ACC-3028 move ContentStoreResolver to domain and EncryptedContentStoreResolver to autoconfigure
