+++ contentgrid-appserver-autoconfigure/src/main/java/com/contentgrid/appserver/autoconfigure/contentstore/FilesystemContentStoreAutoConfiguration.java
@@ -1,6 +1,6 @@
 package com.contentgrid.appserver.autoconfigure.contentstore;
 
-import com.contentgrid.appserver.contentstore.api.resolver.ContentStoreResolver;
+import com.contentgrid.appserver.domain.content.ContentStoreResolver;
 import com.contentgrid.appserver.contentstore.impl.fs.FilesystemContentStore;
 import java.io.IOException;
 import java.nio.file.Files;
