+++ contentgrid-appserver-contentstore-api/build.gradle
@@ -5,6 +5,5 @@ plugins {
 }
 
 dependencies {
-    api project(':contentgrid-appserver-application-model')
 }
